+++ sources/news/articles/uk/england/index.ts
@@ -1,0 +1,2 @@
+export * from "./articles-list";
+export * from "./sources";
