+++ sources/news/articles/uk/national/providers.ts
@@ -1,0 +1,294 @@
+export const bbcNews = {
+	origin: "UK",
+	name: "BBC",
+	description: "BBC News is a British free-to-air television news channel.",
+	url: "https://www.bbc.co.uk",
+	logo: "https://static.files.bbci.co.uk/ws/simorgh-assets/public/news/images/metadata/poster-1024x576.png",
+	rating: 82,
+	leaning: 0,
+};
+
+export const guardian = {
+	origin: "UK",
+	name: "The Guardian",
+	description:
+		"The Guardian is a British news and media website owned by the Guardian Media Group.",
+	url: "https://www.theguardian.com",
+	logo: "https://assets.guim.co.uk/static/frontend/icons/homescreen/apple-touch-icon-512.png",
+	rating: 85,
+	leaning: -0.1,
+};
+export const telegraph = {
+	origin: "UK",
+	name: "The Telegraph",
+	description: "The Telegraph is a British newspaper.",
+	url: "https://www.telegraph.co.uk/",
+	logo: "https://www.telegraph.co.uk/etc.clientlibs/settings/wcm/designs/telegraph/core/clientlibs/core/resources/icons/tmg-share-img.svg",
+	rating: 70,
+	leaning: 0.2,
+};
+
+export const mirror = {
+	origin: "UK",
+	name: "The Mirror",
+	description: "The Mirror is a British national daily tabloid newspaper.",
+	url: "https://www.mirror.co.uk/",
+	logo: "https://s2-prod.mirror.co.uk/@trinitymirrordigital/chameleon-branding/publications/mirror/img/logo-mirror-social-sharing.png",
+	rating: 82,
+	leaning: -0.3,
+};
+
+export const dailyMail = {
+	origin: "UK",
+	name: "The Daily Mail",
+	description: "The Daily Mail is a British daily middle-market newspaper.",
+	url: "https://www.dailymail.co.uk/",
+	logo: "https://www.dailymail.co.uk/static/icons/180.png",
+	rating: 65,
+	leaning: 0.5,
+};
+
+export const metro = {
+	origin: "UK",
+	name: "Metro",
+	description: "Metro is a British free daily newspaper.",
+	url: "https://metro.co.uk/",
+	logo: "https://metro.co.uk/wp-content/client-mu-plugins/metro-branding/images/m-black-share.png",
+	rating: 80,
+	leaning: 0.2,
+};
+
+export const eveningStandard = {
+	origin: "UK",
+	name: "The Evening Standard",
+	description: "The Evening Standard is a local, free daily newspaper.",
+	url: "https://www.standard.co.uk/",
+	logo: "https://www.standard.co.uk/img/shortcut-icons/icon-512x512.png",
+	rating: 80,
+	leaning: 0.3,
+};
+
+export const theSun = {
+	origin: "UK",
+	name: "The Sun",
+	description: "The Sun is a British tabloid newspaper.",
+	url: "https://www.thesun.co.uk/",
+	logo: "https://www.thesun.co.uk/wp-content/uploads/2016/04/icon-e1459786005667.png?strip=all",
+	rating: 65,
+	leaning: 0.5,
+};
+
+export const skyNews = {
+	origin: "UK",
+	name: "Sky News",
+	description: "Sky News is a British free-to-air television news channel.",
+	url: "https://news.sky.com/",
+	logo: "https://news.sky.com/resources/apple-touch-icon.png?v=2",
+	rating: 85,
+	leaning: 0,
+};
+
+export const independent = {
+	origin: "UK",
+	name: "The Independent",
+	description: "The Independent is a British online newspaper.",
+	url: "https://www.independent.co.uk/",
+	logo: "https://www.independent.co.uk/img/shortcut-icons/icon-96x96.png",
+	rating: 85,
+	leaning: -0.1,
+};
+
+export const politicsCoUk = {
+	origin: "UK",
+	name: "politics.co.uk",
+	description: "Politics.co.uk is a British political news website.",
+	url: "https://www.politics.co.uk/",
+	logo: "https://www.politics.co.uk/wp-content/themes/politicsuk/library/images/apple-touch-icon.png",
+	rating: 80,
+	leaning: -0.2,
+};
+
+export const gbNews = {
+	origin: "UK",
+	name: "GB News",
+	description: "GB News is a British news channel and website.",
+	url: "https://www.gbnews.uk/",
+	logo: "https://www.gbnews.com/media-library/eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpbWFnZSI6Imh0dHBzOi8vYXNzZXRzLnJibC5tcy8zMzEwMDg4MS9vcmlnaW4ucG5nIiwiZXhwaXJlc19hdCI6MTc5NzY2ODQ5OX0.yPkj1VxfBk72KU9hNSQwXmOcRlz79vywCKtwhL4rbwA/image.png?width=210",
+	rating: 45,
+	leaning: 0.7,
+};
+export const sheffieldTribune = {
+	origin: "UK",
+	name: "The Sheffield Tribune",
+	description: "The Sheffield Tribune is a British regional daily newspaper.",
+	url: "https://www.sheffieldtribune.co.uk/",
+	logo: "https://www.sheffieldtribune.co.uk/content/images/2024/12/Sheffeildcover02logo2-2.jpg",
+	rating: 80,
+	leaning: 0,
+};
+
+export const theTimes = {
+	origin: "UK",
+	name: "The Times",
+	description: "The Times is a British daily national newspaper.",
+	url: "https://www.thetimes.com/",
+	logo: "https://www.thetimes.com/subscribe/wp-content/uploads/sites/9/2018/04/cropped-T_Favicon_192x1921.png?w=192",
+	rating: 75,
+	leaning: 0.1,
+};
+
+export const ft = {
+	origin: "UK",
+	name: "Financial Times",
+	description: "The Financial Times is a British daily newspaper.",
+	url: "https://www.ft.com/",
+	logo: "https://www.ft.com/__origami/service/image/v2/images/raw/ftlogo-v1%3Abrand-ft-logo-square-coloured?source=update-logos&format=png&width=180&height=180",
+	rating: 75,
+	leaning: 0.1,
+};
+
+export const novara = {
+	origin: "UK",
+	name: "Novara Media",
+	description: "Novara Media is a British independent media company.",
+	url: "https://novaramedia.com/",
+	logo: "https://novaramedia.com/wp-content/uploads/2024/02/OG_Main-Website-1200x630.png",
+	rating: 75,
+	leaning: 0,
+};
+
+export const bylineTimes = {
+	origin: "UK",
+	name: "Byline Times",
+	description: "Byline Times is a British independent media company.",
+	url: "https://bylinetimes.com/",
+	logo: "https://bylinetimes.com/wp-content/uploads/2025/04/Social-Media-Cover-600x314.jpg",
+	rating: 75,
+	leaning: 0,
+};
+
+export const councilEstateMedia = {
+	origin: "UK",
+	name: "Council Estate Media",
+	description: "Council Estate MEdia is a UK substack publication.",
+	url: "https://www.councilestatemedia.uk/",
+	logo: "https://substackcdn.com/image/fetch/e_trim:10:white/e_trim:10:transparent/h_72,c_limit,f_auto,q_auto:good,fl_progressive:steep/https%3A%2F%2Fsubstack-post-media.s3.amazonaws.com%2Fpublic%2Fimages%2F15dd166a-13cb-4fa1-908b-563fca297e21_252x49.jpeg",
+	rating: 75,
+	leaning: 0,
+};
+
+export const owenJones = {
+	origin: "UK",
+	name: "Owen Jones",
+	description: "Owen Jones is a British journalist and author.",
+	url: "https://www.owenjones.news/",
+	logo: "https://substackcdn.com/image/fetch/f_auto,q_auto:good,fl_progressive:steep/https%3A%2F%2Fsubstack-post-media.s3.amazonaws.com%2Fpublic%2Fimages%2F5ea42589-dd0c-455c-9226-a862f0a7c12f%2Fapple-touch-icon-180x180.png",
+	rating: 75,
+	leaning: 0,
+};
+
+export const doubleDownNews = {
+	origin: "UK",
+	name: "Double Down News",
+	description: "Double Down News is a British independent media company.",
+	url: "https://www.doubledown.news/",
+	logo: "http://static1.squarespace.com/static/5b3bed4a75f9ee82a35df852/t/5cc779884e17b6046f5372ca/1556576656282/Cover.jpg?format=1500w",
+	rating: 75,
+	leaning: 0,
+};
+
+export const theEconomist = {
+	origin: "UK",
+	name: "The Economist",
+	description: "The Economist is a UK Economy magazine.",
+	url: "https://www.economist.com/",
+	logo: "https://economist.prod.limio.com/public/8e54554b-dc88-48d8-845d-4cad0e847c00/favicon.f1ea9088.png",
+	rating: 85,
+	leaning: 0,
+};
+export const itv = {
+	origin: "UK",
+	name: "ITV",
+	description: "ITV is a British free-to-air television network.",
+	url: "https://www.itv.com/",
+	logo: "https://app.10ft.itv.com/itvstatic/assets/images/brands/itvx/itvx-logo-for-light-backgrounds.jpg",
+	rating: 85,
+	leaning: 0.1,
+};
+export const channel4 = {
+	origin: "UK",
+	name: "Channel 4",
+	logo: "https://www.channel4.com/favicon-32x32.png",
+	description: "Channel 4 is a British public-service television broadcaster.",
+	url: "https://www.channel4.com/",
+	rating: 85,
+	leaning: 0.1,
+};
+export const channel5 = {
+	origin: "UK",
+	name: "Channel 5",
+	description: "Channel 5 is a British free-to-air television network.",
+	url: "https://www.channel5.com/",
+	logo: "https://www.channel5.com/my5_192.png",
+	rating: 85,
+	leaning: 0.1,
+};
+
+export const theConversation = {
+	origin: "UK",
+	name: "The Conversation",
+	description: "The Conversation is a British independent news website.",
+	url: "https://theconversation.com",
+	logo: "https://images.theconversation.com/files/581997/original/9275d5c0-7dd5-4ddf-b50b-98d537e016f4.png",
+	rating: 90,
+	leaning: 0,
+};
+
+export const thePinkNews = {
+	origin: "UK",
+	name: "The Pink News",
+	description: "The Pink News is a British news website.",
+	url: "https://www.thepinknews.com/",
+	logo: "https://www.thepinknews.com/wp-content/themes/pinknews/.src/images/favicons/mstile-310x310.png",
+	rating: 75,
+	leaning: -0.3,
+};
+
+export const theStar = {
+	origin: "UK",
+	name: "The Star",
+	description: "The Star is a British regional daily newspaper for Sheffield.",
+	url: "https://www.thestar.co.uk/",
+	logo: "https://www.thestar.co.uk/img/placeholder.png",
+	rating: 80,
+	leaning: 0,
+};
+
+export const tldr = {
+	origin: "UK",
+	name: "TLDR",
+	description: "TLDR is a British news website.",
+	url: "https://tldrnews.co.uk/",
+	logo: "https://tldrnews.co.uk/wp-content/uploads/2023/09/cropped-UK-192x192.png",
+	rating: 75,
+	leaning: 0,
+};
+export const declassified = {
+	origin: "UK",
+	name: "Declassified UK",
+	description: "Declassified UK is a British independent media company.",
+	url: "https://declassifieduk.org/",
+	logo: "https://www.declassifieduk.org/wp-content/uploads/2022/10/Declassified_Twitter_Banner_02a.jpg",
+	rating: 75,
+	leaning: 0,
+};
+export const express = {
+	origin: "UK",
+	name: "Daily Express",
+	description:
+		"The Daily Express is a daily national middle-market tabloid newspaper.",
+	url: "https://www.express.co.uk/",
+	logo: "https://cdn.images.express.co.uk/appletouchicons/20160608/apple-touch-icon-180x180.png",
+	rating: 85,
+	leaning: -0.1,
+};
