+++ sources/news/articles/uk/national/sources.ts
@@ -1,0 +1,97 @@
+import {
+	bbcNews,
+	bylineTimes,
+	channel4,
+	channel5,
+	councilEstateMedia,
+	dailyMail,
+	declassified,
+	doubleDownNews,
+	eveningStandard,
+	express,
+	ft,
+	gbNews,
+	guardian,
+	independent,
+	itv,
+	metro,
+	mirror,
+	novara,
+	owenJones,
+	politicsCoUk,
+	sheffieldTribune,
+	skyNews,
+	telegraph,
+	theConversation,
+	theEconomist,
+	thePinkNews,
+	theStar,
+	theSun,
+	theTimes,
+	tldr,
+} from "./providers";
+
+export const ukNationalSources = {
+	BBC: "BBC",
+	GUARDIAN: "The Guardian",
+	TELEGRAPH: "The Telegraph",
+	MIRROR: "The Mirror",
+	DAILY_MAIL: "The Daily Mail",
+	METRO: "Metro",
+	EVENING_STANDARD: "The Evening Standard",
+	SUN: "The Sun",
+	SKY: "Sky News",
+	TIMES: "Times",
+	INDEPENDENT: "The Independent",
+	ITV: "ITV",
+	CHANNEL_4: "Channel 4",
+	CHANNEL_5: "Channel 5",
+	SHEFFIELD_TRIBUNE: "The Sheffield Tribune",
+	THE_ECONOMIST: "The Economist",
+	NOVARA: "Novara Media",
+	BYLINE_TIMES: "Byline Times",
+	COUNCIL_ESTATE_MEDIA: "Council Estate Media",
+	OWEN_JONES: "Owen Jones",
+	DOUBLE_DOWN_NEWS: "Double Down News",
+	THE_CONVERSATION: "The Conversation",
+	THE_PINK_NEWS: "The Pink News",
+	POLITICS_CO_UK: "politics.co.uk",
+	GB_NEWS: "GB News",
+	FT: "FT",
+	THE_STAR: "The Star",
+	DECLASSIFIED: "Declassified UK",
+	TLDR: "TLDR News",
+	EXPRESS: "Daily Express",
+};
+export const ukNationalSourcesMap = new Map([
+	[ukNationalSources.BBC, bbcNews],
+	[ukNationalSources.GUARDIAN, guardian],
+	[ukNationalSources.TELEGRAPH, telegraph],
+	[ukNationalSources.MIRROR, mirror],
+	[ukNationalSources.DAILY_MAIL, dailyMail],
+	[ukNationalSources.METRO, metro],
+	[ukNationalSources.EVENING_STANDARD, eveningStandard],
+	[ukNationalSources.SUN, theSun],
+	[ukNationalSources.SKY, skyNews],
+	[ukNationalSources.INDEPENDENT, independent],
+	[ukNationalSources.ITV, itv],
+	[ukNationalSources.CHANNEL_4, channel4],
+	[ukNationalSources.CHANNEL_5, channel5],
+	[ukNationalSources.SHEFFIELD_TRIBUNE, sheffieldTribune],
+	[ukNationalSources.TIMES, theTimes],
+	[ukNationalSources.THE_ECONOMIST, theEconomist],
+	[ukNationalSources.NOVARA, novara],
+	[ukNationalSources.FT, ft],
+	[ukNationalSources.BYLINE_TIMES, bylineTimes],
+	[ukNationalSources.COUNCIL_ESTATE_MEDIA, councilEstateMedia],
+	[ukNationalSources.OWEN_JONES, owenJones],
+	[ukNationalSources.DOUBLE_DOWN_NEWS, doubleDownNews],
+	[ukNationalSources.POLITICS_CO_UK, politicsCoUk],
+	[ukNationalSources.GB_NEWS, gbNews],
+	[ukNationalSources.THE_CONVERSATION, theConversation],
+	[ukNationalSources.THE_PINK_NEWS, thePinkNews],
+	[ukNationalSources.THE_STAR, theStar],
+	[ukNationalSources.DECLASSIFIED, declassified],
+	[ukNationalSources.TLDR, tldr],
+	[ukNationalSources.EXPRESS, express],
+]);
