+++ sources/news/sources.ts
@@ -15,68 +15,48 @@ import {
 	afp,
 	aljazeera,
 	arabNews,
-	bbcNews,
 	bloomberg,
 	bostonHerald,
 	breitbart,
 	buzzfeed,
-	bylineTimes,
 	cbc,
 	cbs,
-	channel4,
-	channel5,
 	channelNewsAsia,
 	cnbc,
 	cnn,
-	councilEstateMedia,
-	dailyMail,
 	dailyWire,
-	declassiied,
 	democracyNow,
 	denverPost,
-	doubleDownNews,
 	dropSiteNews,
 	dw,
 	eInternationalRelations,
 	electronicIntifada,
 	euronews,
-	eveningStandard,
-	express,
 	firstPost,
 	fox,
 	france24,
-	ft,
-	gbNews,
 	globalIssues,
 	globalNews,
 	globeAndMail,
-	guardian,
 	haaretz,
 	huffpost,
 	ifpNews,
-	independent,
-	itv,
 	japanTimes,
 	kyivIndependent,
 	laTimes,
 	mercuryNews,
-	metro,
 	middleEastEye,
 	mintPressNews,
-	mirror,
 	nbc,
 	ndtv,
 	news24,
 	newYorkPost,
 	newYorkTimes,
-	novara,
 	npr,
-	owenJones,
 	palestineChronicle,
 	palestineNewsNetwork,
 	pbs,
 	politico,
-	politicsCoUk,
 	proPublica,
 	qudsNews,
 	rawStory,
@@ -87,25 +67,20 @@ import {
 	rt,
 	sabc,
 	seatleTimes,
-	sheffieldTribune,
-	skyNews,
 	slate,
 	southChinaMorningPost,
 	sputnik,
 	straitsTimes,
 	sydneyMorningHerald,
-	telegraph,
 	tennesseeHoller,
 	theAmericanConservative,
 	theAtlantic,
 	theBulwark,
 	theCipherBrief,
-	theConversation,
 	theCradle,
 	theDailyBeast,
 	theDailyCaller,
 	theDiplomat,
-	theEconomist,
 	theFederalist,
 	theGrayZone,
 	theHill,
@@ -115,15 +90,10 @@ import {
 	theNewHumanitarian,
 	theNewYorker,
 	theOnion,
-	thePinkNews,
-	theStar,
-	theSun,
-	theTimes,
 	theWorld,
 	time,
 	timesofindia,
 	timesOfIsrael,
-	tldr,
 	torontoStar,
 	trt,
 	truthOut,
@@ -140,23 +110,8 @@ import {
 } from "./providers/providers";
 export const NewsSources = {
 	ARAB_NEWS: "Arab News",
-	BBC: "BBC",
-	GUARDIAN: "The Guardian",
-	TELEGRAPH: "The Telegraph",
-	MIRROR: "The Mirror",
-	DAILY_MAIL: "The Daily Mail",
-	METRO: "Metro",
-	EVENING_STANDARD: "The Evening Standard",
-	SUN: "The Sun",
-	SKY: "Sky News",
-	TIMES: "Times",
-	INDEPENDENT: "The Independent",
+
 	HUFFPOST: "The Huffington Post",
-	POLITICS_CO_UK: "politics.co.uk",
-	THE_CONVERSATION: "The Conversation",
-	THE_PINK_NEWS: "The Pink News",
-	GB_NEWS: "GB News",
-	FT: "FT",
 
 	CNN: "CNN",
 	FOX: "Fox News",
@@ -210,7 +165,7 @@ export const NewsSources = {
 	NDTV: "NDTV",
 	E_INTERNATIONAL_RELATIONS: "E-International Relations",
 	CIPHER_BRIEF: "The Cipher Brief",
-	EXPRESS: "Daily Express",
+
 	EURONEWS: "Euronews",
 	CHANNEL_NEWS_ASIA: "Channel News Asia",
 	NEWS_24: "News24",
@@ -226,10 +181,6 @@ export const NewsSources = {
 	GLOBAL_ISSUES: "Global Issues",
 	UN: "United Nations",
 
-	ITV: "ITV",
-	CHANNEL_4: "Channel 4",
-	CHANNEL_5: "Channel 5",
-
 	NBC: "NBC",
 	PBS: "PBS",
 	REUTERS: "Reuters",
@@ -243,19 +194,10 @@ export const NewsSources = {
 	TRT: "TRT",
 	SABC: "SABC",
 
-	TLDR: "TLDR News",
-
 	//
-	THE_STAR: "The Star",
-	SHEFFIELD_TRIBUNE: "The Sheffield Tribune",
 
-	// Audio
-	THE_ECONOMIST: "The Economist",
-
-	NOVARA: "Novara Media",
 	KYIV_INDEPENDENT: "Kyiv Independent",
 
-	DECLASSIFIED: "Declassified UK",
 	TENNESSEE_HOLLER: "Tennessee Holler",
 
 	THE_INTELLECTUALIST: "The Intellectualist",
@@ -270,12 +212,9 @@ export const NewsSources = {
 	MINT_PRESS_NEWS: "Mint Press News",
 	TRUTHOUT: "Truthout",
 	REVEAL_NEWS: "Reveal News",
-	BYLINE_TIMES: "Byline Times",
-	COUNCIL_ESTATE_MEDIA: "Council Estate Media",
 	DAILY_WIRE: "Daily Wire",
 	SLATE: "Slate",
-	OWEN_JONES: "Owen Jones",
-	DOUBLE_DOWN_NEWS: "Double Down News",
+
 	THE_CRADLE: "The Cradle",
 	ROYA_NEWS: "Roya News",
 	ZETEO: "Zeteo",
@@ -292,21 +231,9 @@ export const NewsSources = {
 // We can type better
 export const newsSourcesMap = new Map<string, any>([
 	[NewsSources.ARAB_NEWS, arabNews],
-	[NewsSources.BBC, bbcNews],
-	[NewsSources.GUARDIAN, guardian],
-	[NewsSources.TELEGRAPH, telegraph],
-	[NewsSources.MIRROR, mirror],
-	[NewsSources.DAILY_MAIL, dailyMail],
-	[NewsSources.METRO, metro],
-	[NewsSources.EVENING_STANDARD, eveningStandard],
-	[NewsSources.SUN, theSun],
-	[NewsSources.SKY, skyNews],
-	[NewsSources.INDEPENDENT, independent],
+
 	[NewsSources.HUFFPOST, huffpost],
-	[NewsSources.POLITICS_CO_UK, politicsCoUk],
-	[NewsSources.THE_CONVERSATION, theConversation],
-	[NewsSources.THE_PINK_NEWS, thePinkNews],
-	[NewsSources.GB_NEWS, gbNews],
+
 	[NewsSources.CNN, cnn],
 	[NewsSources.FOX, fox],
 	[NewsSources.WALL_STREET_JOURNAL, wallStreetJournal],
@@ -358,7 +285,7 @@ export const newsSourcesMap = new Map<string, any>([
 	[NewsSources.NDTV, ndtv],
 	[NewsSources.E_INTERNATIONAL_RELATIONS, eInternationalRelations],
 	[NewsSources.CIPHER_BRIEF, theCipherBrief],
-	[NewsSources.EXPRESS, express],
+
 	[NewsSources.EURONEWS, euronews],
 	[NewsSources.CHANNEL_NEWS_ASIA, channelNewsAsia],
 	[NewsSources.NEWS_24, news24],
@@ -372,9 +299,6 @@ export const newsSourcesMap = new Map<string, any>([
 	[NewsSources.SPUTNIK, sputnik],
 	[NewsSources.GLOBAL_ISSUES, globalIssues],
 	[NewsSources.UN, un],
-	[NewsSources.ITV, itv],
-	[NewsSources.CHANNEL_4, channel4],
-	[NewsSources.CHANNEL_5, channel5],
 	[NewsSources.NBC, nbc],
 	[NewsSources.PBS, pbs],
 	[NewsSources.REUTERS, reuters],
@@ -389,18 +313,8 @@ export const newsSourcesMap = new Map<string, any>([
 	[NewsSources.TRT, trt],
 	[NewsSources.SABC, sabc],
 
-	[NewsSources.THE_STAR, theStar],
-	[NewsSources.SHEFFIELD_TRIBUNE, sheffieldTribune],
-
-	[NewsSources.TIMES, theTimes],
-	[NewsSources.TLDR, tldr],
-
-	[NewsSources.THE_ECONOMIST, theEconomist],
-	[NewsSources.FT, ft],
-
-	[NewsSources.NOVARA, novara],
 	[NewsSources.KYIV_INDEPENDENT, kyivIndependent],
-	[NewsSources.DECLASSIFIED, declassiied],
+
 	[NewsSources.TENNESSEE_HOLLER, tennesseeHoller],
 	[NewsSources.THE_INTELLECTUALIST, theIntellectualist],
 	[NewsSources.DROP_SITE_NEWS, dropSiteNews],
@@ -414,12 +328,8 @@ export const newsSourcesMap = new Map<string, any>([
 	[NewsSources.MINT_PRESS_NEWS, mintPressNews],
 	[NewsSources.TRUTHOUT, truthOut],
 	[NewsSources.REVEAL_NEWS, revealNews],
-	[NewsSources.BYLINE_TIMES, bylineTimes],
-	[NewsSources.COUNCIL_ESTATE_MEDIA, councilEstateMedia],
 	[NewsSources.DAILY_WIRE, dailyWire],
 	[NewsSources.SLATE, slate],
-	[NewsSources.OWEN_JONES, owenJones],
-	[NewsSources.DOUBLE_DOWN_NEWS, doubleDownNews],
 	[NewsSources.THE_CRADLE, theCradle],
 	[NewsSources.ROYA_NEWS, royaNews],
 	[NewsSources.ZETEO, zeteo],
